Read zip skins with one shared reader on every platform
The zip reader was an expect/actual pair: java.util.zip on JVM/Android and a
hand-rolled central-directory walk on iOS, which has no bundled zip facility.
Two implementations of the same format can drift, and only one of them was
under test.

Keep the central-directory reader, move it to commonMain, and delete both
actuals. It only needed a DEFLATE primitive to be portable, so take that from
Okio: its Inflater is a typealias to java.util.zip.Inflater on JVM/Android and
is backed by platform.zlib on iOS, so every target still inflates with native
zlib. Okio was already shipping transitively via Coil, so this adds no new
dependency; kompress, which the iOS reader used for DEFLATE, and its oshi and
karbide transitives are dropped.

Okio's zlib APIs live in a JVM+Native source set. Adding a js or wasm target
would make them invisible to commonMain, so that bridge gets crossed if we
ever come to it.

SkinZipLoaderTest now covers the code every platform runs, rather than the JVM
actual alone.
