--- compose/src/commonJvmAndroidMain/kotlin/warlockfe/warlock3/compose/util/ZipReader.jvmAndroid.kt
@@ -1,19 +1,0 @@
-package warlockfe.warlock3.compose.util
-
-import java.io.ByteArrayInputStream
-import java.util.zip.ZipInputStream
-
-actual fun readZipEntries(bytes: ByteArray): Map<String, ByteArray> {
-    val entries = LinkedHashMap<String, ByteArray>()
-    ZipInputStream(ByteArrayInputStream(bytes)).use { zip ->
-        var entry = zip.nextEntry
-        while (entry != null) {
-            if (!entry.isDirectory) {
-                entries[entry.name] = zip.readBytes()
-            }
-            zip.closeEntry()
-            entry = zip.nextEntry
-        }
-    }
-    return entries
-}
